backend/requirements.txt
# --- Framework / server ---
fastapi==0.116.1
starlette==0.47.2
uvicorn[standard]==0.35.0
gunicorn==21.2.0
a2wsgi==1.10.7

# --- DB ---
SQLAlchemy==2.0.43
PyMySQL==1.1.1
greenlet==3.2.4

# --- Validation / multipart / env ---
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-settings==2.5.2
python-multipart==0.0.20
python-dotenv==1.1.1
email_validator==2.2.0


# --- Vision / face embeddings ---
insightface==0.7.3
onnxruntime==1.19.2
numpy>=2.0.0,<2.3.0
pillow==11.3.0
opencv-python-headless==4.12.0.88

# --- Async / red ---
anyio==4.10.0
h11==0.16.0
idna==3.10
sniffio==1.3.1

# --- Migraciones / plantillas ---
alembic==1.16.4
Mako==1.3.10
MarkupSafe==3.0.2

# --- Crypto / Hashing ---
cryptography==45.0.6
ecdsa==0.19.1
rsa==4.9.1
passlib[bcrypt,argon2]==1.7.4

# --- Seguridad / Rate limiting ---
slowapi==0.1.9
limits==3.13.0

# --- Logging / Utilidades / Testing ---
loguru==0.7.2
click==8.1.8
packaging==24.2
typing_extensions==4.14.1
pytest==8.3.3
tzdata>=2024.1